igad-ai/pyproject.toml
[project]
name = "igad-ai-service"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "boto3",
    "pandas",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "mangum",
    "openai"
]